test(core-utils): properly mock time in time-dependent tests
